Phase C performance: emulator core 2.98x faster (plan C1-C5, C7 part)
Pokemon Red core-only throughput 29.1x -> 86.7x realtime; dmg-acid2
41.3x -> 193.3x. Measured with paired, order-alternated runs
(docs/compatibility/compare.sh) because this machine has fast and slow
states ~15% apart.

C1  Event schedule + absolute clock. New src/schedule.rs; MMU::now.
    Timer/Divider/Serial converted to catch_up(now)/next_event() and to
    storing when the next event is due rather than time since the last.
    MachineCycles widened to u64; its Sub no longer silently saturates.
C2  HALT fast-path (Core::skip_halt). 65% of emulated m-cycles were HALT
    paying full price; skipping to the next scheduled event doubled
    Pokemon on its own. Proved bit-identical to cycle-by-cycle stepping.
C3  Closed-form APU phase timers and noise LFSR.
C4  Mix only when a channel's packed DAC level moves.
C5  Sprite search hoisted out of the pixel loop; per-tile fetch and
    palette instead of per-pixel; per-scanline sprite column mask.
C7  InterruptFlags is a bitmask, so the per-instruction IRQ poll is one
    and plus a trailing_zeros.

Zero fixture regeneration: every representation change kept its
serialised shape behind a Snapshot struct, and the new clock went into
the already-reserved `sched` section.

Verified: default tier 975 passed; slow-tests 1069 passed;
full_playthrough passed to all 8 badges (667s -> 277s); dmg-acid2 and
cgb-acid2 still match their reference PNGs byte for byte; no fixture
drift.

+++ src/pokemon/integration_tests/fixture.rs
@@ -134,7 +134,7 @@ impl TestFixture {
         // rival or an Elite Four room needs far more than the usual threshold — but a fight can also
         // deadlock outright (an attacker out of PP, healing itself forever against a mon it cannot KO),
         // and that has to keep failing fast instead of running to the cycle cap.
-        const BATTLE_STALL_FACTOR: usize = 8;
+        const BATTLE_STALL_FACTOR: u64 = 8;
         let steps = self.agent.policy_steps_remaining();
         let long_running = self.agent.policy_current_step_is_long_running();
         let threshold = if self.agent.in_battle() {
